pyproject.toml
[tool.poetry]
name = "telegramshoter"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.21"
python-decouple = "^3.6"
loguru = "^0.6.0"
validators = "^0.20.0"
pyppeteer = "^1.0.2"
pre-commit = "^2.20.0"
asyncpg = "^0.26.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
